refactor(http): extract reusable server core
