+++ Cargo.toml
@@ -3,6 +3,7 @@
 
 [workspace]
 members = [
+    "lib/http-server",
     "lib/llm",
     "lib/rl",
     "lib/runtime",
@@ -43,6 +44,7 @@ keywords = ["llm", "genai", "inference", "nvidia", "distributed"]
 [workspace.dependencies]
 # Local crates
 dynamo-runtime = { path = "lib/runtime", version = "1.3.0" }
+dynamo-http-server = { path = "lib/http-server", version = "1.3.0" }
 dynamo-llm = { path = "lib/llm", version = "1.3.0" }
 dynamo-rl = { path = "lib/rl", version = "1.3.0" }
 dynamo-tokenizers = { version = "=1.5.0" }
