+++ lib/http-server/src/lib.rs
@@ -1,0 +1,8 @@
+// SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+pub mod disconnect;
+pub mod lifecycle;
+pub mod metrics;
+pub mod request;
+pub mod server;
