+++ lib/llm/src/entrypoint/input/http.rs
@@ -9,7 +9,10 @@ use crate::{
     endpoint_type::EndpointType,
     engines::StreamingEngineAdapter,
     entrypoint::{ChatEngineFactoryCallback, EngineConfig, RouterConfig, input::common},
-    http::service::service_v2::{self, HttpService},
+    http::service::{
+        metrics::update_metrics_from_mdc,
+        service_v2::{self, HttpService},
+    },
     local_model::runtime_config::TokenizerBackend,
     namespace::NamespaceFilter,
     types::openai::{
@@ -17,6 +20,7 @@ use crate::{
         completions::{NvCreateCompletionRequest, NvCreateCompletionResponse},
     },
 };
+use dynamo_http_server::metrics::Metrics;
 use dynamo_runtime::DistributedRuntime;
 use dynamo_runtime::metrics::MetricsHierarchy;
 
@@ -183,7 +187,7 @@ async fn run_watcher(
     migration_max_seq_len: Option<u32>,
     namespace_filter: NamespaceFilter,
     http_service: Arc<HttpService>,
-    metrics: Arc<crate::http::service::metrics::Metrics>,
+    metrics: Arc<Metrics>,
     chat_engine_factory: Option<ChatEngineFactoryCallback>,
     prefill_load_estimator: Option<Arc<dyn dynamo_kv_router::PrefillLoadEstimator>>,
     local_model_path: Option<PathBuf>,
@@ -270,14 +274,11 @@ fn update_http_endpoints(service: Arc<HttpService>, model_type: ModelUpdate) {
 }
 
 /// Updates metrics for model type changes
-fn update_model_metrics(
-    model_type: ModelUpdate,
-    metrics: Arc<crate::http::service::metrics::Metrics>,
-) {
+fn update_model_metrics(model_type: ModelUpdate, metrics: Arc<Metrics>) {
     match model_type {
         ModelUpdate::Added(card) => {
             tracing::debug!("Updating metrics for added model: {}", card.display_name);
-            if let Err(err) = metrics.update_metrics_from_mdc(&card) {
+            if let Err(err) = update_metrics_from_mdc(&metrics, &card) {
                 tracing::warn!(%err, model_name=card.display_name, "update_metrics_from_mdc failed");
             }
         }
