+++ lib/llm/src/http/service.rs
@@ -33,7 +33,6 @@ pub mod realtime;
 pub mod service_v2;
 
 pub use axum;
-pub use metrics::Metrics;
 
 /// Documentation for a route
 #[derive(Debug, Clone)]
