+++ lib/llm/tests/http-service.rs
@@ -3,6 +3,7 @@
 
 use anyhow::Error;
 use async_stream::stream;
+use dynamo_http_server::metrics::{Endpoint, ErrorType, Metrics, RequestType, Status};
 use dynamo_llm::protocols::{
     Annotated,
     codec::SseLineCodec,
@@ -13,12 +14,7 @@ use dynamo_llm::protocols::{
     },
 };
 use dynamo_llm::{
-    http::service::{
-        Metrics,
-        error::HttpError,
-        metrics::{Endpoint, ErrorType, RequestType, Status},
-        service_v2::HttpService,
-    },
+    http::service::{error::HttpError, service_v2::HttpService},
     model_card::ModelDeploymentCard,
 };
 use dynamo_runtime::metrics::prometheus_names::{frontend_service, name_prefix};
