+++ lib/llm/tests/http_metrics.rs
@@ -3,9 +3,10 @@
 
 use anyhow::Error;
 use async_stream::stream;
+use dynamo_http_server::metrics::Endpoint;
 use dynamo_llm::{
     discovery::UNKNOWN_METRIC_MODEL,
-    http::service::{metrics::Endpoint, service_v2::HttpService},
+    http::service::service_v2::HttpService,
     model_card::ModelDeploymentCard,
     preprocessor::LLMMetricAnnotation,
     protocols::{
@@ -604,7 +605,8 @@ mod integration_tests {
     use super::*;
     use dynamo_llm::{
         discovery::ModelWatcher, engines::make_echo_engine, entrypoint::EngineConfig,
-        local_model::LocalModelBuilder, namespace::NamespaceFilter,
+        http::service::metrics::update_metrics_from_mdc, local_model::LocalModelBuilder,
+        namespace::NamespaceFilter,
     };
     use dynamo_runtime::DistributedRuntime;
     use dynamo_runtime::discovery::DiscoveryQuery;
@@ -705,11 +707,7 @@ mod integration_tests {
         // This simulates what the ModelWatcher polling task would do in production
         let _ = manager.save_model_card("test-mdc-key", card.clone());
 
-        if let Err(e) = service
-            .state()
-            .metrics_clone()
-            .update_metrics_from_mdc(&card)
-        {
+        if let Err(e) = update_metrics_from_mdc(&service.state().metrics_clone(), &card) {
             tracing::debug!(
                 model = %card.display_name,
                 error = %e,
